Fixed bug loading wrong resources for resourceless champs.  Also all champ buttons work again.
V0.0.3.  Resourceless champ bug wasnt so much a bug as a design flaw.
Apparently some champs like Aatrox have values for MP and MPRegen
despite not having a resource.  Therefore, testing to see if these
values are 0 is not a good test for determining whether "mp" and
"mpRegen" variables need to be printed to the screen.  Switched instead
to an if-else construct that checks the JSON entry "partype" for mana,
energy, or something else.  The "something else"s have if-else's based
on champ names, if they have an unusual resource like fury or health.
Everyone else (like Garen and Riven) end up in the last else.