app can now retrieve current list of champions with a button
v0.0.7
Got a refresh data button to work which currently only checks if
champion.json exists, and if not, redownloads it.
Most importantly, the app can now retrieve a up-to-date list of all the
champions and have that data available for other processes, like champ
button generation and data lookup.