fixed downloading
v0.0.9
Using AsyncHttpClient libraries with this project was a mess.  Way too powerful for what i needed.  So i threw it out, and just implemented AsyncTask and threading myself.
After several days, the code can again download champion.json and Aatrox.png, with the capability to iteratively download more using looping and a queue class that can be taken in by the AsyncTask.  So glad its finally over.
Next is getting it to download those two files plus three more champ icons, and either version checking system or getting the icon propagating code to use the downloaded icons instead of having the icons compiled in.